requirements.txt
pillow
rembg
onnxruntime
tqdm
black
